+++ packages/catlog/src/stdlib/analyses/ode/mass_action/v0/mass_action.rs
