+++ packages/catlog/src/stdlib/analyses/ode/mass_action/v0/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod mass_action;
