+++ packages/catlog/src/stdlib/analyses/ode/polynomial_ode/v0/polynomial_ode.rs
