--- packages/catlog/src/stdlib/analyses/ode/v0/mod.rs
@@ -1,6 +1,0 @@
-//! Version 0 analyses, i.e. those that have been since modified (to v1) with breaking changes.
-
-pub mod linear_ode;
-pub mod lotka_volterra;
-pub mod mass_action;
-pub mod polynomial_ode;
