--- packages/catlog/src/stdlib/analyses/ode/v1/mod.rs
@@ -1,7 +1,0 @@
-//! Version 1 analyses, i.e. those that have received breaking changes since v0.
-
-pub mod linear_ode;
-pub mod lotka_volterra;
-pub mod mass_action;
-pub mod migrate;
-pub mod polynomial_ode;
